@Jasper-Bekkers 2021-11-08
My 2 cents, Microsoft is actively asking for feedback so some of this may be useful to share: https://github.com/microsoft/windows-rs/issues/1285