@kennykerr 2021-11-08
Nice! 

Please create new issues for any feedback: https://github.com/microsoft/windows-rs/issues

Very happy to clean up any warts you may find. Compile time will improve, as the win32 metadata namespaces are being refactored as we speak. That is a high priority for me.

If it helps, here's another reference port that may be interesting to compare: https://github.com/yoshuawuyts/miow/compare/kenny?expand=1