@maroider 2021-11-09
I must admit that I'm still on the fence when it comes to adopting the `windows` crate. `winapi`'s build times are already longer than I'd prefer, and I doubt windows-rs can match `winapi`'s built times when it seemingly generates a whole bunch of convenience wrappers on top of the raw bindings.

That's not to say that I don't appreciate that you're taking the time to do this, it's just that I'm not quite sure the tradeoffs (as things currently stand) are worth it in the end.