@clemenswasser 2021-11-09
> I must admit that I'm still on the fence when it comes to adopting the `windows` crate. `winapi`'s build times are already longer than I'd prefer, and I doubt windows-rs can match `winapi`'s built times when it seemingly generates a whole bunch of convenience wrappers on top of the raw bindings.
> 
> That's not to say that I don't appreciate that you're taking the time to do this, it's just that I'm not quite sure the tradeoffs (as things currently stand) are worth it in the end.

I would argue, winapi is currently unmaintained, so nothing on the compile times or other areas will significantly change. windows-rs on the other hand is a pretty young project with Microsoft behind it, compile times will get better (maybe even better than winapi, who knows?). What I want to say is that windows-rs already has many significant improvements today, with more to come.