@ChrisDenton 2021-11-09
Just to be clear, `winapi` isn't unmaintained. It's not merging feature PRs (e.g. new APIs) but it is doing so for bug fixes. However, you're right that means nothing has significantly changed in a long time. There is apparently going to be a call for additional maintainers soon so that situation may change.

On the other hand (as noted in the OP) `windows-rs` has recently switched to using pre-built binaries and should (soon-ish) have a `no-std` option. Both of these may reduce its build times if that's the main concern. Admittedly it would be ideal to test project build times to see how they compare. One advantage `windows-rs` may have is that it has some finer grained namespaces whereas `winapi` tends to namespaced by entire header files (which may themselves depend on other headers, and so on).