@MarijnS95 2021-11-09
@maroider @ChrisDenton IMO windows-rs is the future:

- Officially built and backed by Microsoft, so expect proper and timely updates;
- Already much, much better/safer APIs, even if there's still quite a way to go;
- Completely autogenerated, APIs will stay up to date as long as the generator is updated to support (probably ever-expanding) win32metadata - and generator improvements affect all API surface;
- Afaik intended to officially replace winapi in the first place.

I don't think it makes sense for `winapi` to live on, besides the occasional bugfix to support older crates that have not yet or are in the process of migrating.

As for the open issues (eg. compile speed), that's a bit of a chicken-and-egg problem. The faster windows-rs gets widespread adoption, the quicker issues can be properly triaged and addressed.

However I can also see that it's quite early and with windows-rs breaking updates coming in hot, `winit` will have to move fast in tracking and updating `windows-rs` in order to not fall behind or run into the age-old duplicate crate "issue" (affecting compile times and cargo-deny if used, and public API if at all exposed anywhere).