@MarijnS95 2021-11-10
@clemenswasser Fyi windows-rs just released 0.26 with slightly refactored namespaces. I hope that's not causing many issues here as they're mostly related to Direct3D :)