@clemenswasser 2021-11-10
Not using the `std` feature saves ~10 seconds for me and saves a few mallocs.