@clemenswasser 2021-11-11
> > Not using the `std` feature saves ~10 seconds of build time for me and a few mallocs.
> 
> We're getting hangs internally with [6629e52](https://github.com/rust-windowing/winit/commit/6629e52a7462e128122fd9078b67b1591e68e28f) (`no_std`), before that everything is fine. Haven't bisected the root cause yet.

Does 693fef2 fix it? I forgot to add the null terminators, sorry 😅