@MarijnS95 2021-11-11
> Does [693fef2](https://github.com/rust-windowing/winit/commit/693fef296afc30ad8700e5ac225fd545b105a803) fix it? I forgot to add the null terminators, sorry sweat_smile

Yeah I completely forgot that I reviewed that when being alerted this morning that our upgrade to windows-rs 0.26 and this PR may have caused the breakage. Surprisingly no segfaults, but I've been told that it's fixed now with the null-terminators added back!