@kennykerr 2021-11-12
> I must admit that I'm still on the fence when it comes to adopting the `windows` crate. `winapi`'s build times are already longer than I'd prefer, and I doubt windows-rs can match `winapi`'s built times when it seemingly generates a whole bunch of convenience wrappers on top of the raw bindings.

The `windows-sys` crate is coming soon ([rationale](https://github.com/microsoft/windows-rs/pull/1310)). It provides only low-level declarations and may be more suitable for cases like this. Here's a [simple comparison of compile time](https://github.com/microsoft/windows-rs/pull/1314).