@maroider 2021-11-13
> The `windows-sys` crate is coming soon ([rationale](https://github.com/microsoft/windows-rs/pull/1310)). It provides only low-level declarations and may be more suitable for cases like this. Here's a [simple comparison of compile time](https://github.com/microsoft/windows-rs/pull/1314).

This seems like it would address my biggest concern with migrating to windows-rs.

I do, of course, have smaller complaints, like the atypical module name style and the deepening of `use` statements, but these are hardly showstoppers.

I also have a minor concern about the soundness of marking simple wrapper-structs like `LRESULT(isize)` as `repr(C)` rather than `repr(transparent)` when said struct is used directly when defining `extern fn`s. The latter more accurately captures the intent of "this struct should be laid out in memory as if it was the contained type".