@kennykerr 2021-11-16
FYI

The windows-sys crate has now been published: https://crates.io/crates/windows-sys

Here's a small windowing example: https://github.com/microsoft/windows-samples-rs/tree/master/create_window_sys