@kennykerr 2021-11-16
> I also have a minor concern about the soundness of marking simple wrapper-structs like LRESULT(isize) as repr(C) rather than repr(transparent) when said struct is used directly when defining extern fns. The latter more accurately captures the intent of "this struct should be laid out in memory as if it was the contained type".

The `windows` and `windows-sys` crates are derived from metadata and that metadata distinguishes between `C` and `transparent` layouts. For example here's `LRESULT` using `repr(transparent)`:

https://github.com/microsoft/windows-rs/blob/master/crates/deps/sys/src/Windows/Win32/Foundation/mod.rs#L2021-L2022

While `MSG` uses `repr(C)`:

https://github.com/microsoft/windows-rs/blob/master/crates/deps/sys/src/Windows/Win32/UI/WindowsAndMessaging/mod.rs#L2795-L2804