@clemenswasser 2022-02-25
Some compile time benchmarks:

current master: b7e7755eddddc375273ffa28dfd1d5bdc86bee9f  with winapi 0.3.9:
```
> hyperfine.exe -n 10 --warmup 2 -p 'cargo clean' 'cargo build'
Benchmark 1: 10
  Time (mean ± σ):      3.641 s ±  0.033 s    [User: 0.000 s, System: 0.003 s]
  Range (min … max):    3.610 s …  3.722 s    10 runs
```

this branch: a48c1e77ebca784d97f7c1b47cbe8c29fe17fb77 with windows-sys 0.33 & 0.32 (through parking-lot):
```
> hyperfine.exe -n 10 --warmup 2 -p 'cargo clean' 'cargo build'
Benchmark 1: 10
  Time (mean ± σ):      3.910 s ±  0.017 s    [User: 0.000 s, System: 0.000 s]
  Range (min … max):    3.872 s …  3.938 s    10 runs
```

this branch: a48c1e77ebca784d97f7c1b47cbe8c29fe17fb77  with windows-sys 0.33 and custom parking-lot also updated to windows-sys 0.33:
```
> hyperfine.exe -n 10 --warmup 2 -p 'cargo clean' 'cargo build'
Benchmark 1: 10
  Time (mean ± σ):      3.970 s ±  0.034 s    [User: 0.000 s, System: 0.000 s]
  Range (min … max):    3.933 s …  4.041 s    10 runs
```

So currently a light regression.